+++ backend/api/db.py
@@ -1,0 +1,8 @@
+from backend.database import ObjectDatabase
+from backend.database import Database
+
+
+database = ObjectDatabase()
+
+def get_db() -> Database:
+    return database
